Commit for all the the work of the week:
[SPI/LCD] I tried hard to  make the writing on the lcd screen of
the mylab faster (trying using interrupts). I could not manage to
have something working.

[GPU] Passed the bootrom, nothing was displayed on the screen so
I am trying to figure out why. I made a function in the main file
that display all the tiles contained in the tiledata space (0x8000).
Everything is working (I can see all the tiles of Tetris rom) so
if there is a problem, it is in the tilemap space (0x9800).